Unify own memcpy/memset/memcmp
Currently, we have a few implementations of memcpy/memset/memcmp
functions:

 1. Ones that comes from a standard library (glibc/musl/whatever).
 2. Own implementation used by criu/pie/, named builtin_mem*
 3. (Optional) arch-optimized versions of above
    (memcpy for x86, memcpy and memcmp for power).
 4. Own implementation provided/used by compel plugins, named std_*

There are two reasons why 3-5 exist:
 a) code which is not linked to a libc (i.e. parasite/restorer and
    also compel plugins).
 b) C compiler, which might generate calls to memcpy, memset, memcmp,
    memmove calls as it seem fit (so far we haven't seen memmove
    being used by either gcc or clang on any arch supported)

So, having own mem* functions is required. What's messed up is
how many implementations and different names for those we have.

This commit is an attempt to untie the knot described.

A few implementation considerations:

First, let's stick to just one implementation, and have it in compel
std plugin.

Second, let's use the code from criu as it comes with arch-optimized
versions. Surely this means the patch is bigger.

Third, let's call the functions memcpy/memset/memcmp (i.e. no std_ or
builtin_ prefix, no aliases), as this is required by the compiler
(see "b" above) and we are not using libc in this code anyway, so no
names conflict.

Notes:

 - FORTIFY_SOURCE had to be disabled for compel/plugins as in Ubuntu
   Trusty x86_64 it plays tricks with standard includes, replacing
   calls to memcpy() with __memcpy_chk().

 - builtin_strncmp() had to be moved to compel std plugin as
   criu/include/asm-generic/string.h is gone. Currently the only user
   is util-vdso.c

Signed-off-by: Kir Kolyshkin <kir@openvz.org>