python: futurize
Make python code work with both python2 and python3.
Initially I tried to do that manually but later I found
this wonderful tool called futurize.

This patch is a result of running `futurize --stage1 -w`
and `futurize --stage2 -w`.

As a result, I can run it with either python2 or python3.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>